refactor: keyring-internal-api/src/{btc,eth,sol} -> keyring-api/src
