+++ packages/keyring-api/src/api/balance.ts
