+++ packages/keyring-api/src/api/keyring.ts
