+++ packages/keyring-api/src/btc/types.ts
@@ -1,8 +1,9 @@
-import { KeyringAccountStruct, BtcAccountType } from '@metamask/keyring-api';
 import { object } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
 import { string, array, enums, refine, literal } from '@metamask/superstruct';
 import { bech32 } from 'bech32';
+
+import { BtcAccountType, KeyringAccountStruct } from '../api';
 
 export const BtcP2wpkhAddressStruct = refine(
   string(),
