+++ packages/keyring-api/src/eth/index.ts
@@ -1,0 +1,3 @@
+export * from './erc4337';
+export * from './types';
+export * from './utils';
