+++ packages/keyring-api/src/eth/types.ts
@@ -1,7 +1,8 @@
-import { EthAccountType, KeyringAccountStruct } from '@metamask/keyring-api';
 import { object, definePattern } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
 import { array, enums, literal } from '@metamask/superstruct';
+
+import { EthAccountType, KeyringAccountStruct } from '../api';
 
 export const EthBytesStruct = definePattern('EthBytes', /^0x[0-9a-f]*$/iu);
