+++ packages/keyring-api/src/index.ts
@@ -1,8 +1,7 @@
-export * from './account';
-export * from './balance';
-export * from './caip';
-export * from './export';
-export type * from './keyring';
-export * from './request';
-export * from './response';
+export * from './api';
+export * from './btc';
+export * from './sol';
+export * from './eth';
 export type * from './contexts';
+export * from './rpc';
+export * from './events';
