+++ packages/keyring-api/src/sol/types.ts
@@ -1,7 +1,8 @@
-import { KeyringAccountStruct, SolAccountType } from '@metamask/keyring-api';
 import { object, definePattern } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
 import { array, enums, literal } from '@metamask/superstruct';
+
+import { KeyringAccountStruct, SolAccountType } from '../api';
 
 /**
  * Solana addresses are represented in the format of a 256-bit ed25519 public key and
