+++ packages/keyring-internal-api/src/eth/index.ts
@@ -1,4 +1,1 @@
-export * from './erc4337';
-export * from './types';
-export * from './utils';
 export type * from './EthKeyring';
