+++ packages/keyring-internal-api/src/index.ts
@@ -1,7 +1,3 @@
-export * from './btc';
-export * from './eth';
-export * from './sol';
+export type * from './eth';
 export * from './api';
-export * from './events';
-export * from './rpc';
 export * from './types';
