+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -6,19 +6,6 @@ import type { TypedTransaction } from '@ethereumjs/tx';
 import { TransactionFactory } from '@ethereumjs/tx';
 import type { TypedDataV1, TypedMessage } from '@metamask/eth-sig-util';
 import { SignTypedDataVersion } from '@metamask/eth-sig-util';
-import type {
-  KeyringAccount,
-  KeyringExecutionContext,
-  KeyringResponse,
-} from '@metamask/keyring-api';
-import type {
-  BtcMethod,
-  EthBaseTransaction,
-  EthBaseUserOperation,
-  EthUserOperation,
-  EthUserOperationPatch,
-  InternalAccount,
-} from '@metamask/keyring-internal-api';
 import {
   AccountCreatedEventStruct,
   AccountDeletedEventStruct,
@@ -31,7 +18,18 @@ import {
   KeyringEvent,
   RequestApprovedEventStruct,
   RequestRejectedEventStruct,
-} from '@metamask/keyring-internal-api';
+} from '@metamask/keyring-api';
+import type {
+  KeyringAccount,
+  KeyringExecutionContext,
+  KeyringResponse,
+  BtcMethod,
+  EthBaseTransaction,
+  EthBaseUserOperation,
+  EthUserOperation,
+  EthUserOperationPatch,
+} from '@metamask/keyring-api';
+import type { InternalAccount } from '@metamask/keyring-internal-api';
 import { KeyringSnapControllerClient } from '@metamask/keyring-snap-client';
 import type { SnapController } from '@metamask/snaps-controllers';
 import type { SnapId } from '@metamask/snaps-sdk';
