+++ packages/keyring-snap-client/src/KeyringClient.ts
@@ -7,6 +7,7 @@ import type {
   CaipAssetType,
   Balance,
 } from '@metamask/keyring-api';
+import { KeyringRpcMethod } from '@metamask/keyring-api';
 import {
   ApproveRequestResponseStruct,
   CreateAccountResponseStruct,
@@ -21,7 +22,6 @@ import {
   RejectRequestResponseStruct,
   SubmitRequestResponseStruct,
   UpdateAccountResponseStruct,
-  KeyringRpcMethod,
 } from '@metamask/keyring-internal-api';
 import type { JsonRpcRequest } from '@metamask/keyring-snap-sdk';
 import { strictMask } from '@metamask/keyring-utils';
