--- packages/keyring-snap-sdk/src/events.ts
@@ -1,1 +1,0 @@
-export { KeyringEvent } from '@metamask/keyring-internal-api';
