+++ packages/keyring-snap-sdk/src/index.ts
@@ -1,4 +1,3 @@
 export * from './JsonRpcRequest';
-export * from './events';
 export * from './rpc-handler';
 export * from './snap-utils';
