+++ packages/keyring-snap-sdk/src/snap-utils.ts
@@ -1,7 +1,6 @@
+import type { KeyringEvent } from '@metamask/keyring-api';
 import type { SnapsProvider } from '@metamask/snaps-sdk';
 import type { Json } from '@metamask/utils';
-
-import type { KeyringEvent } from './events';
 
 /**
  * Emit a keyring event from a snap.
